+++ src/slskd/Bootstrap/VirtualSoulfindServiceCollectionExtensions.cs
@@ -48,7 +48,6 @@ public static class VirtualSoulfindServiceCollectionExtensions
         services.AddSingleton<VirtualSoulfind.ShadowIndex.IShadowIndexQuery, VirtualSoulfind.ShadowIndex.ShadowIndexQuery>();
         services.AddSingleton<VirtualSoulfind.ShadowIndex.IShardMerger, VirtualSoulfind.ShadowIndex.ShardMerger>();
         services.AddSingleton<VirtualSoulfind.ShadowIndex.IShardCache, VirtualSoulfind.ShadowIndex.ShardCache>();
-        services.AddSingleton<VirtualSoulfind.ShadowIndex.IDhtRateLimiter, VirtualSoulfind.ShadowIndex.DhtRateLimiter>();
         services.AddSingleton<VirtualSoulfind.Scenes.ISceneService, VirtualSoulfind.Scenes.SceneService>();
         services.AddSingleton<VirtualSoulfind.Scenes.ISceneAnnouncementService>(sp =>
             new VirtualSoulfind.Scenes.SceneAnnouncementService(